share_it_website/backend/requirements.txt
flask
oracledb
python-dotenv
bcrypt
flask-cors
